Rename peer_v2 to peer
